fix recommendations component, move component in folder page, and fix crud for history
