+++ src/components/movies/cardMovieItem.tsx
@@ -51,7 +51,6 @@ const CardMovieItem = ({ product }: CardMovieItemProps) => {
 
   const handleDelete = () => {
     if (product?.id) {
-      dispatch(command.products.delete(product?.id));
       dispatch(command.history.delete(product?.id));
       navigate('/movies');
     }
