+++ src/components/movies/nowPlaying.tsx
